+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 ### Version 13.1
+- Dashboard pie color is retained in local storage
 
 ### Version 13.0
 - Fixed ingredients CSV redeclarations